# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Five 网页验证版接口
  version: 1.0.0
  summary: 公开网页、单人后台与后端共同使用的 P0 接口契约
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    本文件是 Five P0 接口字段、状态、错误码和版本保护规则的唯一事实源。
    `contentVersion`、各种 ID 和修订号都是不透明值，调用方不得解析其字符串。
    P0 只有网页端，因此接口不包含 `platform`、`sourcePlatform` 或 `targetPlatform`。
    `channelId` 只用于区分链接分发来源，不得改变日期、档位、正文或图片。
    所有 `/admin/api/v1` 响应（包括错误）必须使用 `Cache-Control: no-store`，防止退出后从缓存读取后台内容或核对凭证。
servers:
  - url: /
    description: 与网页同源的后端
...
security: []
paths:
...
  /admin/api/v1/daily-content-drafts/{draftId}/image-assets:
    get:
      operationId: listDailyContentDraftImageAssets
      tags: [Admin images]
      summary: 读取草稿下已上传的图片候选与后台预览地址
      security:
        - AdminSession: []
      parameters:
        - $ref: "#/components/parameters/DraftId"
      responses:
        "200":
          description: 返回该草稿当前的图片候选
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
            ETag:
              $ref: "#/components/headers/DraftETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftImageAssetList"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
    post:
      operationId: uploadDailyContentDraftImageAsset
      tags: [Admin images]
      summary: 上传一张绑定草稿与命理日的图片
      description: |
        文件与 JSON 元数据使用两个 multipart part；metadata 必须是没有 filename 的普通表单字段，
        字段值是符合 ImageAssetUploadMetadata 的 UTF-8 JSON 文本。服务端计算文件校验值、尺寸和格式；
        imageSlot 是新上传请求的必填字段；本次人工上传会原子成为该槽位的显式选择。
        历史未分槽记录只允许通过原幂等键重放，不接受新的未分槽上传。
        用户访问时不会触发生图。网络重试必须复用同一个 `Idempotency-Key`。
      security:
        - AdminSession: []
      parameters:
        - $ref: "#/components/parameters/DraftId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/XCsrfToken"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ImageAssetUploadMultipartRequest"
            examples:
              aiGenerated:
                value:
                  file: "(binary image bytes)"
                  imageSlot: required_primary
                  metadata:
                    altText: 红色上衣和绿色下装的通勤穿搭
                    sourceType: ai_generated
                    generationMethod: codex
                    declaredModel: gpt-image-2
                    promptVersion: outfit-prompt-v3
                    generatedAt: "2026-07-31T09:00:00+08:00"
                    reproductionReference: generation-job-01
                    sourceMaterialReferences: [formula-triple-01]
                    rightsRecordIds: [rights-ai-terms-01]
                    aiLabelStatus: pending
            encoding:
              file:
                contentType: image/avif, image/webp, image/jpeg, image/png
              metadata:
                contentType: application/json
      responses:
        "201":
          description: 图片已上传并增加草稿修订号
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
            ETag:
              $ref: "#/components/headers/DraftETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftImageAssetResult"
        "400":
          $ref: "#/components/responses/ImageUploadInvalid"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/CsrfValidationFailed"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/RevisionMismatch"
        "413":
          $ref: "#/components/responses/ImageFileTooLarge"
        "415":
          $ref: "#/components/responses/ImageMediaTypeUnsupported"
        "422":
          $ref: "#/components/responses/ImageReviewIncomplete"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
  /admin/api/v1/daily-content-drafts/{draftId}/image-assets/{assetId}/review:
    post:
      operationId: reviewDailyContentDraftImageAsset
      tags: [Admin images]
      summary: 逐图登记结构化人工检查
      description: |
        批准时六项检查必须全部通过，权利状态必须已清理，AI 素材还必须完成适用标识。
        审核记录追加保存审核编号、维护者账号和服务端时间。复制自冻结版本且
        `reviewLocked` 为 true 的素材不能在当前草稿重新审核；需要调整时必须上传新素材。
      security:
        - AdminSession: []
      parameters:
        - $ref: "#/components/parameters/DraftId"
        - $ref: "#/components/parameters/AssetId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/XCsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageAssetReviewRequest"
            examples:
              approved:
                value:
                  decision: approved
                  notes: 手机与微信预览清楚，颜色与配方一致。
                  rightsStatus: cleared
                  aiLabelStatus: complete
                  colorAndCopyConsistency: passed
                  garmentAndPersonIntegrity: passed
                  rightsAndIdentityRisk: passed
                  scenarioAndImitability: passed
                  mobileAndWechatPreview: passed
                  aiLabelCompliance: passed
      responses:
        "200":
          description: 人工检查已保存并增加草稿修订号
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
            ETag:
              $ref: "#/components/headers/DraftETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftImageAssetResult"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/CsrfValidationFailed"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/RevisionMismatch"
        "422":
          $ref: "#/components/responses/ImageReviewIncomplete"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
  /admin/api/v1/image-assets/{assetId}/preview:
    get:
      operationId: previewAdminImageAsset
      tags: [Admin images]
      summary: 在后台预览尚未公开或已经冻结的图片文件
      description: |
        仅供已认证维护者检查手机与微信显示效果；响应不进入公共缓存，也不能替代公开图片 URL。
        文件是否允许公开仍由内容状态与当前图片交付投影决定。
      security:
        - AdminSession: []
      parameters:
        - $ref: "#/components/parameters/AssetId"
      responses:
        "200":
          description: 返回原始上传文件
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
            X-Content-Type-Options:
              schema:
                type: string
                const: nosniff
          content:
            image/avif:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
...
  /admin/api/v1/daily-content-versions/{contentVersion}/daily-image-set:
    get:
      operationId: getDailyContentVersionImageSet
      tags: [Admin images]
      summary: 查看不可变图片引用与当前追加式交付投影
      description: |
        2+1 只统计封面槽位，细节图和降级素材另算。响应只在已认证后台返回内部来源、
        权利和人工检查记录，不得复用为公共响应。
      security:
        - AdminSession: []
      parameters:
        - $ref: "#/components/parameters/ContentVersionPath"
      responses:
        "200":
          description: 返回该内容版本的每日图片组
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
            ETag:
              $ref: "#/components/headers/LifecycleETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDailyImageSet"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
  /admin/api/v1/daily-content-versions/{contentVersion}/image-assets/{assetId}/withdraw:
    post:
      operationId: withdrawDailyContentVersionImageAsset
      tags: [Admin images]
      summary: 追加单图下线事件并更新图片交付投影
      description: |
        不修改不可变内容快照。必备封面切换到同一快照内审核通过的降级素材；可选封面
        或问题细节图从公开交付中省略。重复请求返回第一次写入的同一结果。
      security:
        - AdminSession: []
      parameters:
        - $ref: "#/components/parameters/ContentVersionPath"
        - $ref: "#/components/parameters/AssetId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/XCsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawImageAssetRequest"
            examples:
              copyrightRisk:
                value:
                  expectedActiveContentVersion: fd-20260715-r3
                  reason: 发现来源材料的授权范围不包含公开传播，立即停止交付。
      responses:
        "200":
          description: 单图下线事件、审计和新交付投影已原子保存
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
            ETag:
              $ref: "#/components/headers/LifecycleETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageAssetWithdrawalResult"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/CsrfValidationFailed"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/RevisionMismatch"
        "422":
          $ref: "#/components/responses/ImageWithdrawalBlocked"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
...
  /admin/api/v1/daily-content-drafts/{draftId}/image-assets/{assetId}/selection:
    post:
      operationId: selectDailyContentDraftImageAsset
      tags: [Admin images]
      summary: 显式选择一个候选作为草稿槽位素材
      description: |
        使用草稿修订号与幂等键原子更新槽位选择；同槽旧候选仍保留但不参与自动冻结。
        冻结只读取显式选择，不按上传时间、数组顺序或提示词版本猜测。
      security:
        - AdminSession: []
      parameters:
        - $ref: "#/components/parameters/DraftId"
        - $ref: "#/components/parameters/AssetId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/XCsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectDraftImageAssetRequest"
      responses:
        "200":
          description: 候选已成为指定槽位的显式选择，并增加草稿修订号
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
            ETag:
              $ref: "#/components/headers/DraftETag"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftImageAssetResult"
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/CsrfValidationFailed"
        "404":
          $ref: "#/components/responses/ResourceNotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "412":
          $ref: "#/components/responses/RevisionMismatch"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
...
components:
  securitySchemes:
    AdminSession:
      type: apiKey
      in: cookie
      name: five_admin_session
      description: |
        Issue #31 已确认的同源安全会话。生产 Cookie 固定为 `HttpOnly`、`Secure`、
        `SameSite=Strict`、`Path=/admin` 且无 `Domain`；闲置 30 分钟、绝对 12 小时失效。
        所有后台写操作还必须校验可信同源 `Origin` 与会话绑定的 `X-CSRF-Token`。
  parameters:
...
    DraftId:
      name: draftId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/OpaqueId"
    AssetId:
      name: assetId
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/OpaqueId"
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: |
        草稿写操作使用响应中的草稿 ETag；生命周期写操作使用内容版本 ETag；
        全局紧急控制使用公开内容开关 ETag。调用方必须原样回传，不能自行拼接。
      schema:
        type: string
        minLength: 3
        maxLength: 256
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: 同一业务意图的网络重试必须复用同一个高熵不透明值。
      schema:
        type: string
        minLength: 16
        maxLength: 128
        pattern: '^[-A-Za-z0-9_:.]{16,128}$'
    XCsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: |
        当前后台会话绑定的随机 CSRF 令牌。后台写操作还必须由服务端校验可信同源 `Origin`；
        令牌只能保存在运行中页面内存，不能放入 localStorage、URL 或日志。
      schema:
        type: string
        minLength: 32
        maxLength: 256
  headers:
...
    XRequestId:
      description: 当前外部请求的追踪编号；缓存命中时也必须按请求生成。
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
    DraftETag:
      description: 草稿当前修订号的强 ETag，例如 `"draft:7"`。
      schema:
        type: string
        example: '"draft:7"'
...
    NoStoreCacheControl:
      description: 认证、恢复、安全记录和紧急控制响应不得被浏览器或中间层保存。
      schema:
        type: string
        const: no-store
...
  schemas:
...
    DraftRevision:
      type: integer
      minimum: 1
...
    ImageAssetUploadMultipartRequest:
      type: object
      additionalProperties: false
      required: [file, imageSlot, metadata]
      properties:
        file:
          type: string
          format: binary
        imageSlot:
          $ref: "#/components/schemas/DailyImageSlot"
          description: 必填。明确把本次人工上传候选选入对应槽位。
        metadata:
          $ref: "#/components/schemas/ImageAssetUploadMetadata"
          description: 在线路中序列化为没有 filename 的普通 multipart 字段；字段值为 UTF-8 JSON 文本。
...
    DraftImageAssetResult:
      type: object
      additionalProperties: false
      required:
        [draftId, fortuneDate, draftRevision, imageSlot, selectedForSlot, asset, previewUrl, reviewLocked]
      properties:
        draftId:
          $ref: "#/components/schemas/OpaqueId"
        fortuneDate:
          $ref: "#/components/schemas/FortuneDate"
        draftRevision:
          $ref: "#/components/schemas/DraftRevision"
        imageSlot:
          oneOf:
            - $ref: "#/components/schemas/DailyImageSlot"
            - type: "null"
          description: 候选在当前草稿中的目标槽位；旧历史候选可为空，服务端不得按上传顺序猜测。
        selectedForSlot:
          type: boolean
          description: 仅当该候选是其 imageSlot 当前显式选择时为 true；未分槽候选恒为 false。
        asset:
          $ref: "#/components/schemas/AdminImageAsset"
        previewUrl:
          $ref: "#/components/schemas/AdminImagePreviewUrl"
        reviewLocked:
          type: boolean
          description: 为 true 时素材来自已冻结快照，不允许在当前草稿内重新审核；如需调整必须上传新素材。
...
    DraftImageAssetList:
      type: object
      additionalProperties: false
      required: [draftId, fortuneDate, draftRevision, items]
      properties:
        draftId:
          $ref: "#/components/schemas/OpaqueId"
        fortuneDate:
          $ref: "#/components/schemas/FortuneDate"
        draftRevision:
          $ref: "#/components/schemas/DraftRevision"
        items:
          type: array
          items:
            $ref: "#/components/schemas/DraftImageCandidate"
...
  responses:
...
    ImageUploadInvalid:
      description: multipart 结构、图片文件或 JSON 元数据无效（`IMAGE_FILE_INVALID` 或 `INVALID_ARGUMENT`）
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    Unauthenticated:
      description: 后台会话不存在或已失效
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    CsrfValidationFailed:
      description: 后台写请求的可信同源 `Origin` 或会话绑定 CSRF 令牌校验失败
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    Conflict:
      description: 幂等键、活跃版本或状态发生冲突
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ResourceNotFound:
      description: 后台资源或任务不存在
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    RevisionMismatch:
      description: "`If-Match` 不是资源当前修订号"
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        ETag:
          description: 当前资源 ETag，调用方必须刷新后再操作。
          schema:
            type: string
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          examples:
            stale:
              $ref: "#/components/examples/RevisionMismatchError"
...